+++ packages/@livestore/common/src/derived-mutations.ts
@@ -1,7 +1,7 @@
-import type { SqliteDsl } from '@livestore/db-schema'
 import type { GetValForKey } from '@livestore/utils'
 import { ReadonlyRecord, Schema } from '@livestore/utils/effect'
 
+import type { SqliteDsl } from './schema/db-schema/mod.js'
 import type * as MutationEvent from './schema/MutationEvent.js'
 import { defineMutation } from './schema/mutations.js'
 import { getDefaultValuesDecoded } from './schema/schema-helpers.js'