+++ packages/@livestore/common/src/schema/schema-helpers.ts
@@ -1,7 +1,7 @@
-import { SqliteDsl } from '@livestore/db-schema'
 import { shouldNeverHappen } from '@livestore/utils'
 import { pipe, ReadonlyRecord, Schema } from '@livestore/utils/effect'
 
+import { SqliteDsl } from '../schema/db-schema/mod.js'
 import type { TableDef, TableDefBase } from './table-def.js'
 
 export const getDefaultValuesEncoded = <TTableDef extends TableDef>(